agent/requirements.txt
requests
python-dotenv
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-google-genai
openai
